requirements.txt
transformers>=4.0.0,<=4.21.3
torch>=1.3.0,<=1.12.1
sentencepiece
jsonnet
tqdm
uniplot
networkx==2.5
sacremoses
openpyxl
xlsxwriter
pyxlsb
streamlit
st-annotated-text
pyxlsb
xlsxwriter
pandas
pytz